tests/pyproject.toml
[project]
name = "integration-tests"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pytest>=8.0",
    "requests>=2.31",
    "paho-mqtt>=2.1.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
